Convert volume fields to decimal; expand column sizes
Change BoxVolume and LiquidVolume from BIGINT/long to DECIMAL(18,2) across the project: update Database/Setup.sql, EF model mappings in ApplicationDbContext, and UploadRecord (long? -> decimal?). Increase max lengths for Barcode (100 -> 200), StoreId (32 -> 64), StoreName (200 -> 400), ErrorDescription (500 -> 1000), and TransactionType (20 -> 40). Remove unnecessary casts in FileBuilder formatting for the volume fields. Run a DB migration or apply the SQL changes to align the schema with the model and preserve data.

+++ Models/UploadRecord.cs
@@ -9,8 +9,8 @@ public class UploadRecord
     public decimal? Width { get; set; }
     public decimal? Height { get; set; }
     public decimal? Weight { get; set; }
-    public long? BoxVolume { get; set; }
-    public long? LiquidVolume { get; set; }
+    public decimal? BoxVolume { get; set; }
+    public decimal? LiquidVolume { get; set; }
     public bool? NoDimension { get; set; }
     public bool? NoWeight { get; set; }
     public bool? Sent { get; set; }
